added tests and implemented eqv for string for passing tests
