+++ tests/r6rs.scm
@@ -72,7 +72,6 @@
            42)
 
 ;; 1.11 Continuations
-
 (assert-eq (+ 1 (call-with-current-continuation
                  (lambda (escape)
                    (+ 2 (escape 3)))))
